Update patch set 2
Patch Set 2: Code-Review-1

Although I did not test it:

This is not in any way secure. Native code will always go around the limitation (eg. I guess Firefox will not be blocked)
As the code is running is running in the context of the checking application, it is also possible to bypass the restriction from Java code.

We should not provide a security/privacy feature, that is known to be insecure (although Xprivacy did so).
To do this the right way, one has to use iptables, that will affect everything and is secure (or atleast not easily exploitable)

Patch-set: 2
Reviewer: Gerrit User 1001175 <1001175@56841c99-7c94-43d4-bf0b-cb4707e66d55>
Label: Code-Review=-1